modified:   backend/analytics/__pycache__/calculadora_score.cpython-312.pyc 	modified:   backend/analytics/__pycache__/comparativos.cpython-312.pyc 	modified:   backend/analytics/calculadora_score.py 	modified:   backend/analytics/comparativos.py 	new file:   pages/3_Analise_Receita.py 	modified:   views/__pycache__/vis_analise.cpython-312.pyc 	new file:   views/__pycache__/vis_receita.cpython-312.pyc 	modified:   views/components/__pycache__/glossary.cpython-312.pyc 	modified:   views/components/__pycache__/metrics.cpython-312.pyc 	modified:   views/components/glossary.py 	modified:   views/components/metrics.py 	modified:   views/vis_analise.py 	new file:   views/vis_receita.py

+++ views/vis_analise.py
@@ -172,7 +172,7 @@ def render_analise(df_mestre):
     # Tabela 2: Ranking Geral
     # Aqui usamos o Rank_Geral real que calculamos, pois na tabela geral queremos saber a posição absoluta
     df_view_geral = df_score.sort_values('Power_Score', ascending=False).copy()
-    df_view_geral['#'] = df_view_geral['Rank_Geral']
+    df_view_geral['#'] = df_view_geral['Rank_Geral'].astype(int)
     
     render_ranking_table(
         df_view_geral,
